==> notebooks/0_data_describe.ipynb <==
## 0. Setting

import pandas as pd
import os
import geopandas as gpd
import numpy as np
from tqdm import tqdm
import time
import matplotlib.pyplot as plt
import random
import string

# set working directory
os.chdir("D:/DATA/KEDVAT")


# 1. count missing data

VATRPTD103 = pd.read_csv("VATRPTD0103.txt", sep='|', header=None)
VATTXDT0108 = pd.read_csv("VATTXDT0108.txt", sep='|', header=None)
VATTXSR0103 = pd.read_csv("VATTXSR0103.txt", sep='|', header=None)

chunk_size = 100000000
VATTXDT0204 = pd.read_csv("VATTXDT0204.txt", sep = '|', header=None, chunksize=chunk_size)

df_dict = {"VATRPTD103": VATRPTD103, "VATTXDT0108" : VATTXDT0108, "VATTXSR0103" : VATTXSR0103, "VATTXDT0204" : VATTXDT0204}

for key in df_dict.keys():
    if key == "VATTXDT0204":
        missing_total = None
        for chunk in tqdm(VATTXDT0204, desc = f'chunk in VATTXDT0204'):
            chunk_missing = chunk.isna().sum()
            if missing_total is None:
                missing_total =  chunk_missing
            else:
                missing_total += chunk_missing

        result_df = missing_total.reset_index()
        result_df.to_csv(f"C:/Users/user/Documents/김용기/{key}_missing.csv")
    
    data = df_dict[key]
    print(f'length: {len(data)}')
    data.isna().sum().reset_index().to_csv(f"C:/Users/user/Documents/김용기/{key}_missing.csv")

# 2. mock data

n_rows = 10000
VATRPTD103 = pd.read_csv("VATRPTD0103.txt", sep='|', header=None, nrows = n_rows)
VATRPTD103.columns = ['1_BZNO', '2_BASE_YY', '3_PRD_CCD', '4_CONO_PID', '5_KEDCD', '6_CO_BZNO_CCD', '7_TX_PRD_FROM', '8_TX_PRD_TO', '9_TX_PRD_SUTB_YN', '10_RPTD_CCD', '11_RGSIM_CLS', '12_TTBIL_DLV_AM', '13_BOA_RPT_CR_CARD_CSH_RCPT_AM', '14_BOA_RPT_CR_CARD_CSH_RCPT_AM', '15_TETC_AM', '16_NTAX_TBIL_AM', '17_NTAX_ETC_AM', '18_DIT_EXP_AM_RANGE', '19_VAT_ZTAX_APL_SUM_RANGE', '20_PR_AM', '21_BADB_TICDR_TA', '22_TSTD_SAM', '23_BIL_DLV_AM', '24_SALE_T_AM', '25_GNR_PAM', '26_FASET_PAM', '27_PR_BUY_AM', '28_BUY_BUY_JA_ISS_AM', '29_EDDC_BUY_AM', '30_BUY_AM_SUM', '31_DDC_NAM', '32_DUC_SUM_AM', '33_BIL_RCV_AM', '34_BUY_T_AM', '35_BZNO_SIMSID']

VATTXDT0108 = pd.read_csv("VATTXDT0108.txt", sep='|', header=None, nrows = n_rows)
VATTXDT0108.columns = ['1_BZNO', '2_BASE_YY', '3_PRD_CCD', '4_SALE_BUY_CCD', '5_VAT_ITM_KCD', '6_VAT_SEQ', '7_CONO_PID', '8_KEDCD', '9_CO_BZNO_CCD', '10_TX_PRD_FROM', '11_TX_PRD_TO ', '12_TX_PRD_SUTB_YN', '13_TXPL_TXRI_RNK', '14_TXRI', '15_TXPL_BZNO', '16_TXPL_PID', '17_TXPL_KEDCD', '18_TXPL_NM', '19_TXPL_TCN', '20_TXPL_TSPPR', '21_TXPL_TTAX_AM', '22_BZNO_SIMSID', '23_TXPL_BZNO_SIMSID']

VATTXSR0103 = pd.read_csv("VATTXSR0103.txt", sep='|', header=None, nrows = n_rows)
VATTXSR0103.columns = ['1_BZNO', '2_BASE_YY', '3_PRD_CCD', '4_SALE_BUY_CCD', '5_VAT_ITM_KCD', '6_CONO_PID', '7_KEDCD', '8_CO_BZNO_CCD', '9_TX_PRD_FROM', '10_TX_PRD_TO ', '11_TX_PRD_SUTB_YN', '12_TTXPL_CN', '13_TXPL_CN_BZNO', '14_TXPL_CN_BZNO', '15_TXPL_CN_PID', '16_TXPL_CN_KEDCD_ISS_CN', '17_TCN', '18_CN_BZNO', '19_CN_PID', '20_CN_KEDCD_ISS_CN', '21_TSPPR', '22_SPPR_BZNO', '23_SPPR_PID', '24_SPPR_KEDCD_ISS_CN', '25_TTAX_AM', '26_TAX_AM_BZNO', '27_TAX_AM_PID', '28_TAX_AM_KEDCD_ISS_CN', '29_BZNO_SIMSID']

VATTXDT0204 = pd.read_csv("VATTXDT0204.txt", sep = '|', header=None, nrows = n_rows)
VATTXDT0204.columns = ['1_BZNO', '2_BASE_YY', '3_PRD_CCD', '4_SALE_BUY_CCD', '5_VAT_ITM_KCD', '6_VAT_SEQ', '7_CONO_PID', '8_KEDCD', '9_CO_BZNO_CCD', '10_TX_PRD_FROM', '11_TX_PRD_TO ', '12_TX_PRD_SUTB_YN', '13_TXPL_TXRI_RNK', '14_TXRI', '15_BF_BZC_CD', '16_BZC_CD', '17_TXPL_TCN', '18_TXPL_TSPPR', '19_TXPL_TTAX_AM', '20_BZNO_SIMSID']

df_dict = {"VATRPTD103": VATRPTD103, "VATTXDT0108" : VATTXDT0108, "VATTXSR0103" : VATTXSR0103, "VATTXDT0204" : VATTXDT0204}

# unique values for each columns
for key in df_dict.keys():
    print(f'\n\n{key}')
    df = df_dict[key]
    for col in df.columns:
        print(f'{col}: {len(df[col].unique())}')

# 0108sample
os.makedirs("C:/Users/user/Documents/김용기/2025/temp", exist_ok=True)

for ratio in [0.00001, 0.001]:
    skip = lambda i: 1 > 0 and (random.random() > ratio)

    VATTXDT0108_sampled = pd.read_csv("VATTXDT0108.txt", sep='|', header=None, skiprows= skip)
    VATTXDT0108_sampled.to_csv(f"C:/Users/user/Documents/김용기/2025/temp/VATTXDT0108_sampled_{ratio:.3%}.csv", index=False)

for ratio in [0.00001, 0.001]: 
    df = pd.read_csv(f"C:/Users/user/Documents/김용기/2025/temp/VATTXDT0108_sampled_{ratio:.3%}.csv")
    df.columns = ['1_BZNO', '2_BASE_YY', '3_PRD_CCD', '4_SALE_BUY_CCD', '5_VAT_ITM_KCD', '6_VAT_SEQ', '7_CONO_PID', '8_KEDCD', '9_CO_BZNO_CCD', '10_TX_PRD_FROM', '11_TX_PRD_TO ', '12_TX_PRD_SUTB_YN', '13_TXPL_TXRI_RNK', '14_TXRI', '15_TXPL_BZNO', '16_TXPL_PID', '17_TXPL_KEDCD', '18_TXPL_NM', '19_TXPL_TCN', '20_TXPL_TSPPR', '21_TXPL_TTAX_AM', '22_BZNO_SIMSID', '23_TXPL_BZNO_SIMSID']

    #'1_BZNO', '15_TXPL_BZNO'
    cols = ['1_BZNO', '15_TXPL_BZNO']
    for col in cols:
        uniques = df[col].unique()
        def random_string():
            return ''.join(random.choices(string.digits, k = 10))
        randoms = [random_string()[:5]+ '*****' for _ in uniques]
        code_dict = dict(zip(uniques, randoms))
        df[col] = df[col].map(code_dict)
    del randoms, code_dict

    #'22_BZNO_SIMSID', '23_TXPL_BZNO_SIMSID'
    def random_string():
        return ''.join(random.choices(string.ascii_letters + string.digits, k =10))
    uniques = list(set(np.concatenate([df['1_BZNO'], df['15_TXPL_BZNO']])))

    randoms = [random_string() for _ in uniques]
    code_dict = dict(zip(uniques, randoms))

        # randomly replace with None for '15_TXPL_BZNO'
    uniques_TXPL = list(set(df['15_TXPL_BZNO']) - set(df['1_BZNO']))
    eligible = [k for k in uniques_TXPL if k in code_dict]
    n = len(df['15_TXPL_BZNO'].unique())
    k = int(n*0.47)
    to_None = random.sample(uniques_TXPL, k = min(k, len(eligible)))
    for key in to_None:
        code_dict[key] = None

    df['22_BZNO_SIMSID'] = df['1_BZNO'].map(code_dict)
    df['23_TXPL_BZNO_SIMSID'] = df['15_TXPL_BZNO'].map(code_dict)
    del randoms, code_dict, to_None


    #'2_BASE_YY', '4_SALE_BUY_CCD', 
    cols = ['2_BASE_YY', '4_SALE_BUY_CCD', '5_VAT_ITM_KCD', '9_CO_BZNO_CCD', '6_VAT_SEQ', '13_TXPL_TXRI_RNK', '14_TXRI']
    for col in cols:
        randoms = df[col].copy()
        random.shuffle(randoms)
        df[col] = randoms
    del randoms


    #'3_PRD_CCD', '10_TX_PRD_FROM', '11_TX_PRD_TO ', '12_TX_PRD_SUTB_YN'
    cols = ['3_PRD_CCD', '10_TX_PRD_FROM', '11_TX_PRD_TO ', '12_TX_PRD_SUTB_YN']
    randoms = list(df[cols].itertuples(index=False, name=None))
    random.shuffle(randoms)
    unzipped = list(zip(*randoms))
    for i, col in enumerate(cols):
        df[col] = unzipped[i]
    del randoms, unzipped


    #'19_TXPL_TCN', '20_TXPL_TSPPR', '21_TXPL_TTAX_AM'
    cols = ['19_TXPL_TCN', '20_TXPL_TSPPR', '21_TXPL_TTAX_AM']
    randoms = list(df[cols].itertuples(index=False, name=None))
    random.shuffle(randoms)
    unzipped = list(zip(*randoms))
    for i, col in enumerate(cols):
        df[col] = unzipped[i]
    del randoms, unzipped

    df.to_csv(f"C:/Users/user/Documents/김용기/2025/temp/VATTXDT0108_sampled_{ratio:.3%}_mock.csv", index=False)
